% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{chpexp}
\alias{chpexp}
\title{Cumulative hazard function of the PE distribution}
\usage{
chpexp(x, tgrid, rates)
}
\arguments{
\item{x}{vector of time points.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates.}
}
\value{
Cumulative hazard function of the PE distribution.
}
\description{
Cumulative hazard function of the PE distribution
}
