% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{dpexp}
\alias{dpexp}
\title{Cumulative distribution function of the PE distribution}
\usage{
dpexp(x, tgrid, rates, log = FALSE)
}
\arguments{
\item{x}{vector of time points.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Density function of the PE distribution.
}
\description{
Cumulative distribution function of the PE distribution
}
