% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGrid}
\alias{getGrid}
\title{Computes the time grid from the auxiliary vector U.}
\usage{
getGrid(U, ftgrid)
}
\arguments{
\item{U}{vector of change point indicators}

\item{ftgrid}{vector with the finest time grid (distinct observed failure times)}
}
\value{
the time grid associated with the auxiliary vector U.
}
\description{
Computes the time grid from the auxiliary vector U.
}
