% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{hpexp}
\alias{hpexp}
\title{Hazard function of the PE distribution}
\usage{
hpexp(x, tgrid, rates)
}
\arguments{
\item{x}{vector of time points.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates.}
}
\value{
Hazard function of the PE distribution.
}
\description{
Hazard function of the PE distribution
}
