% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{ppexp}
\alias{ppexp}
\title{Cumulative distribution function of the PE distribution}
\usage{
ppexp(x, tgrid, rates, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{vector of time points.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates}

\item{lower.tail}{logical; if TRUE (default), probabilities are $P[X <= x]$, otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Cumulative distribution function of the PE distribution.
}
\description{
Cumulative distribution function of the PE distribution
}
