% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{qpexp}
\alias{qpexp}
\title{Quantile of the PE distribution}
\usage{
qpexp(p, tgrid, rates, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of quantiles.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates}

\item{lower.tail}{logical; if TRUE (default), probabilities are $P[X <= x]$, otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Cumulative distribution function of the PE distribution.
}
\description{
Quantile of the PE distribution
}
