% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexp.R
\name{rpexp}
\alias{rpexp}
\title{Random generation of the PE distribution}
\usage{
rpexp(n, tgrid, rates)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{tgrid}{vector of time grid knots.}

\item{rates}{vector of failure rates}
}
\value{
Sample of size n of the PE distribution.
}
\description{
Random generation of the PE distribution
}
